add/clear rules feature, refactoring
