--- plugins/honeypot_file_plugin.py
@@ -1,57 +1,0 @@
-import argparse
-import re
-import sys
-
-from core.plugin import Plugin
-from config import PathConfig
-
-VERSION = "0.01"
-PLUGIN_NAME = "Filesystem"
-
-def check_path_arg(value: str) -> str:
-    """
-    TODO: better validation, more options of valid paths (relative, absolute, ~/...)
-    :param value:
-    :return
-    """
-    valid = re.search('^(/[^/\0]+)+/?$', value)
-    if not valid:
-        raise argparse.ArgumentTypeError(value  + " is not a valid path")
-    return value
-
-
-class FileSystemPlugin(Plugin):
-    @staticmethod
-    def init_args_parser(subparser):
-        parser = subparser.add_parser(PLUGIN_NAME.lower())
-        parser.set_defaults(func=FileSystemPlugin.handle_command)
-        group = parser.add_mutually_exclusive_group(required=True)
-        group.add_argument('-v', '--version', help='View currently installed version.', action='version', version=f"{PLUGIN_NAME} v{VERSION}")
-        group.add_argument('-af', '--add-file', type=check_path_arg, help='Add a honeypot file rule.')
-        group.add_argument('-ad', '--add-directory', type=check_path_arg, help='Add a honeypot directory rule.')
-        group.add_argument('-r', '--remove', type=check_path_arg, help='Remove a honeypot filesystem rule.')
-        group.add_argument('-c', '--clear', help='Clear the honeypot file rules.', action="store_true")
-
-    @staticmethod
-    def handle_command(args: argparse.Namespace):
-        if args.add_file:
-            rule = f"-a always,exit -F arch=b64 -F path={args.add} -F perm=rw -k honeypot_file"
-            separator = ''
-            rule_exists = False
-            with open(PathConfig.RULES_FILE_PATH, 'r') as rules_file:
-                rules = rules_file.read()
-                if rules.__len__() > 0:
-                    separator = '\n'
-                if rule in rules:
-                    rule_exists = True
-
-            if rule_exists:
-                print("rule already exists.")
-                sys.exit(1)
-
-            with open(PathConfig.RULES_FILE_PATH, 'a') as rules_file:
-                rules_file.write(separator + rule)
-
-        if args.clear:
-            with open(PathConfig.RULES_FILE_PATH, 'w') as log_file:
-                log_file.write('')
